Fix missing ns3 namespace compilation error
